+++ src/main/java/project/backend/business/post/PostService.java
@@ -16,9 +16,9 @@ import project.backend.business.post.request.CreatePostServiceRequest;
 import project.backend.business.post.request.PostDetailServiceRequest;
 import project.backend.business.post.request.PostListServiceRequest;
 import project.backend.business.post.response.CreateUpdatePostResponse;
-import project.backend.business.post.response.PostDetailDto;
+import project.backend.business.post.response.dto.PostDetailDto;
 import project.backend.business.post.response.PostDetailResponse;
-import project.backend.business.post.response.PostListDto;
+import project.backend.business.post.response.dto.PostListDto;
 import project.backend.business.post.response.PostListResponse;
 import project.backend.business.user.implement.UserReader;
 import project.backend.common.error.CustomException;
