+++ src/main/java/project/backend/business/post/implement/PostReader.java
@@ -9,8 +9,8 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
 import project.backend.business.common.DateTimeManager;
 import project.backend.business.post.request.PostDetailServiceRequest;
-import project.backend.business.post.response.PostDetailDto;
-import project.backend.business.post.response.PostListDto;
+import project.backend.business.post.response.dto.PostDetailDto;
+import project.backend.business.post.response.dto.PostListDto;
 import project.backend.business.tag.implement.TagReader;
 import project.backend.common.error.CustomException;
 import project.backend.common.error.ErrorCode;
