+++ src/main/java/project/backend/business/post/response/PostListResponse.java
@@ -3,6 +3,7 @@ package project.backend.business.post.response;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
+import project.backend.business.post.response.dto.PostListDto;
 
 @Getter
 public class PostListResponse {
