+++ src/main/java/project/backend/business/post/response/dto/PostListDto.java
@@ -1,4 +1,4 @@
-package project.backend.business.post.response;
+package project.backend.business.post.response.dto;
 
 import java.util.List;
 import lombok.Builder;
