+++ src/main/java/project/backend/presentation/user/UserController.java
@@ -1,4 +1,4 @@
-package project.backend.presentation.user.controller;
+package project.backend.presentation.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import project.backend.business.user.UserService;
 import project.backend.business.user.response.UserInfoResponse;
+import project.backend.presentation.user.docs.UserControllerDocs;
 import project.backend.security.aop.AssignCurrentUserInfo;
 import project.backend.security.aop.CurrentUserInfo;
